Technical errors and game plan for next year Part 1
Next year I am going to need all the help for my game because what is happening is that I don't have more time and with the small amount for time I get help, it is becoming harder for me to get all of my work done for the game.
So, I am going to need all of the help I need in order to get this done and hopefully I can gat my future started out by winning the competition.